pyproject.toml
[tool.pytest.ini_options]
testpaths = ["tests/python"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dart-cesm-interface"
version = "0.1.0"
description = "DART-CESM Interface Components"
authors = [
    {name = "DART Development Team"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = []

[tool.setuptools.packages.find]
include = ["cime_config*", "param_templates*"]
exclude = ["tests*", "DART*", "nuopc_driver*", "cesm_build_templates*"]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "f90nml>=1.4",
    "pyyaml>=6.0"
]